Development

See merge request abhi-shah/north-trend-oms!659

+++ resources/views/profile/index.blade.php
@@ -282,7 +282,8 @@
                                     @else
                                     <tr>
                                       <th> <b>Over Credit Limit:</b> </th>
-                                      <td>{{(number_format($data->current_account_balance + @$data->open_orders_balance) - @$data->credit_limit)}}</td>
+                                      <td>
+                                        {{number_format((float)(($data->current_account_balance + @$data->open_orders_balance) - @$data->credit_limit))}}</td>
                                     </tr>
                                     @endif
                                   </thead>
